🔨  added some svg and adapt some features

+++ resources/views/components/button/button.blade.php
@@ -1,9 +1,11 @@
 @props(['icon'=>null])
+@props(['slot' => null])
 <button type="submit" {{ $attributes->class(['buttons p-2.5 flex-row flex-auto justify-center items-center gap-2 flex
-nav_item_hover
-'])  }}>
+nav_item_hover'])  }}>
     @if(!is_null($icon))
         <x-dynamic-component :component="'icon.'.$icon"/>
     @endif
-    <span class="truncate ">{{$slot}}</span>
+    @if(!is_null($slot))
+        <span class="truncate ">{{$slot}}</span>
+    @endif
 </button>
